Refactor: svg 파일 하나로 통합
